Protocols: Update k-point sampling protocols
Based on more extensive tests that also involved a broad range of smearings,
a new set of protocols were devised:

| Name       | Smearing (Ry) | k-points distance (1/A) |
|------------|---------------|-------------------------|
| Precision  |        0.0125 |                    0.10 |
| Efficiency |        0.0200 |                    0.15 |
| Fast       |        0.0275 |                    0.30 |

Here we already update the values in the `PwBaseWorkChain` protocol file. In
a future PR it would be nice to set up a system for protocol aliasing, so
we can also change the names without breaking protocol methods downstream.