Fix maximum number of iterations reached
In the current implementation of the `PwRelaxWorkChain`, an error is raised if the maximum number
of iterations is reached. This check is located in the `should_run_relax` method. In case this error is raised,
it is ignored as the method is called in a `while_` loop which expects a
boolean argument
(https://github.com/aiidateam/plumpy/blob/master/src/plumpy/workchains.py#L395-L399).
Therefore, the statement is moved to the `inspect_relax` method.